pyproject.toml
[project]
name = "kodo"
version = "0.4.0"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk",
    "python-dotenv>=1.0",
    "simple-term-menu",
    "pydantic-ai[anthropic,google]",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["kodo*"]

[project.scripts]
kodo = "kodo.cli:main"
